Add EntityInfo component
